+++ realm-browser/src/main/java/com/dd/realmbrowser/RealmFilesActivity.java
@@ -3,6 +3,7 @@ package com.dd.realmbrowser;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.AdapterView;
@@ -21,7 +22,7 @@ public class RealmFilesActivity extends AppCompatActivity {
     private List<String> mIgnoreExtensionList;
     private ArrayAdapter<String> mAdapter;
 
-    public static void start(Activity activity) {
+    public static void start(@NonNull Activity activity) {
         Intent intent = new Intent(activity, RealmFilesActivity.class);
         activity.startActivity(intent);
     }
@@ -32,8 +33,8 @@ public class RealmFilesActivity extends AppCompatActivity {
         setContentView(R.layout.ac_realm_list_view);
 
         mIgnoreExtensionList = new ArrayList<>();
-        mIgnoreExtensionList.add("log");
-        mIgnoreExtensionList.add("lock");
+        mIgnoreExtensionList.add(".log");
+        mIgnoreExtensionList.add(".lock");
 
         File dataDir = new File(getApplicationInfo().dataDir, "files");
         File[] files = dataDir.listFiles();
@@ -61,7 +62,7 @@ public class RealmFilesActivity extends AppCompatActivity {
         int index = fileName.lastIndexOf(".");
         if (index > 0) {
             String extension = fileName.substring(index);
-            isValid = mIgnoreExtensionList.contains(extension);
+            isValid = !mIgnoreExtensionList.contains(extension);
         }
         return isValid;
     }