+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.2
+
+Enchantments
+* #7 Display notification to launch Realm Browser
+
+Fixed issues
+* #9 Not show data from another module (different package name)
+
 ## 0.0.1
 
 * Initial release